Removed identify() call from AnalyticsIdentifier
Renamed AnalyticsIdentifier* -> AnalyticsPage (as now it only does client page view tracking)
